Move updateWrapper from IceGrowth
I replaced the overElements call to updateWrapper in IceGrowth with a
call to iLateral->update(tst). The idea is to separate out the logic of
lateral ice growth and melt out of IceGrowth itself and into
ILateralIceSpread modules. This does require ILateralIceSpread.hpp to
define and share three variables; for change in ice concentration,
ice thickness, and snow thickness.
