+++ physics/src/IceGrowth.cpp
@@ -1,7 +1,7 @@
 /*!
  * @file IceGrowth.cpp
  *
- * @date 02 May 2025
+ * @date 04 Jun 2025
  * @author Tim Spain <timothy.spain@nersc.no>
  * @author Einar Ólason <einar.olason@nersc.no>
  */
@@ -28,25 +28,15 @@ IceGrowth::IceGrowth()
     , hsnow(ModelArray::Type::H)
     , hice0(ModelArray::Type::H)
     , hsnow0(ModelArray::Type::H)
-    , newice(ModelArray::Type::H)
-    , deltaCIce(ModelArray::Type::H)
-    , deltaCFreeze(ModelArray::Type::H)
-    , deltaCMelt(ModelArray::Type::H)
     , hIceCell(getStore())
     , hSnowCell(getStore())
     , cice0(getStore())
     , qow(getStore())
-    , mixedLayerBulkHeatCapacity(getStore())
-    , sst(getStore())
-    , tf(getStore())
     , deltaHi(getStore())
 {
     getStore().registerArray(Shared::H_ICE, &hice, RW);
     getStore().registerArray(Shared::C_ICE, &cice, RW);
     getStore().registerArray(Shared::H_SNOW, &hsnow, RW);
-    getStore().registerArray(Shared::NEW_ICE, &newice, RW);
-    getStore().registerArray(Shared::HSNOW_MELT, &snowMelt, RW);
-    getStore().registerArray(Shared::DELTA_CICE, &deltaCIce, RW);
 
     getStore().registerArray(Protected::HTRUE_ICE, &hice0, RO);
     getStore().registerArray(Protected::HTRUE_SNOW, &hsnow0, RO);
@@ -63,11 +53,6 @@ void IceGrowth::setData(const ModelState::DataMap& ms)
     hsnow.resize();
     hice0.resize();
     hsnow0.resize();
-    newice.resize();
-    snowMelt.resize();
-    deltaCFreeze.resize();
-    deltaCMelt.resize();
-    deltaCIce.resize();
 }
 
 ModelState IceGrowth::getStateDiagnostic() const
@@ -155,17 +140,13 @@ void IceGrowth::update(const TimestepTime& tsTime)
 {
     // Copy the ice data from the prognostic fields to the modifiable fields.
     initializeThicknesses();
-    overElements([this](size_t i, const TimestepTime& tst) { this->applyLimits(i, tst); }, tsTime);
 
     // The snowMelt array is not currently filled with data, but it used elsewhere
     // FIXME calculate a true value for snowMelt
-    snowMelt = 0;
 
     if (doThermo) {
         iVertical->update(tsTime);
-        // new ice formation
-        overElements(
-            [this](size_t i, const TimestepTime& tst) { this->updateWrapper(i, tst); }, tsTime);
+        iLateral->update(tsTime);
     }
 
     // Damage always heals, even if there's no active thermo
@@ -185,8 +166,6 @@ void IceGrowth::initializeThicknesses()
 // but only if ice concentration is non-zero.
 void IceGrowth::initializeThicknessesElement(size_t i, const TimestepTime&)
 {
-    deltaCIce[i] = 0;
-
     if (cice0[i] > 0 && hIceCell[i] > 0) {
         hice[i] = hice0[i] = hIceCell[i] / cice0[i];
         hsnow[i] = hsnow0[i] = hSnowCell[i] / cice0[i];
@@ -194,78 +173,6 @@ void IceGrowth::initializeThicknessesElement(size_t i, const TimestepTime&)
         hice[i] = hice0[i] = 0.;
         hsnow[i] = hsnow0[i] = 0.;
         cice[i] = 0.;
-    }
-
-    // reset the new ice volume array
-    newice[i] = 0;
-}
-
-void IceGrowth::newIceFormation(size_t i, const TimestepTime& tst)
-{
-    // Flux cooling the ocean from open water
-    // TODO Add assimilation fluxes here
-    double coolingFlux = qow[i];
-    // Temperature change of the mixed layer during this timestep
-    double deltaTml = -coolingFlux / mixedLayerBulkHeatCapacity[i] * tst.step;
-    // Initial temperature
-    double t0 = sst[i];
-    // Freezing point temperature
-    double tf0 = tf[i];
-    // Final temperature
-    double t1 = t0 + deltaTml;
-
-    // deal with cooling below the freezing point
-    if (t1 < tf0) {
-        // Heat lost cooling the mixed layer to freezing point
-        double sensibleFlux = (tf0 - t0) / deltaTml * coolingFlux;
-        // Any heat beyond that is latent heat forming new ice
-        double latentFlux = coolingFlux - sensibleFlux;
-
-        qow[i] = sensibleFlux;
-        newice[i] = latentFlux * tst.step * (1 - cice[i]) / (Ice::Lf * Ice::rho);
-    } else {
-        newice[i] = 0;
-    }
-}
-
-// Update thickness with concentration
-static double updateThickness(double& thick, double newConc, double deltaC, double deltaV)
-{
-    return thick += (deltaV - thick * deltaC) / newConc;
-}
-
-void IceGrowth::lateralIceSpread(size_t i, const TimestepTime& tstep)
-{
-    deltaCMelt[i] = 0;
-    deltaCFreeze[i] = 0;
-    iLateral->freeze(
-        tstep, hice[i], hsnow[i], deltaHi[i], newice[i], cice[i], qow[i], deltaCFreeze[i]);
-    if (deltaHi[i] < 0) {
-        // Note that the cell-averaged hice0 is converted to a ice averaged value
-        iLateral->melt(tstep, hice0[i], hsnow[i], deltaHi[i], cice[i], qow[i], deltaCMelt[i]);
-    }
-    deltaCIce[i] = deltaCFreeze[i] + deltaCMelt[i];
-    cice[i] = (hice[i] > 0 || newice[i] > 0) ? cice[i] + deltaCIce[i] : 0;
-    if (cice[i] >= IceMinima::c()) {
-        // The updated ice thickness must conserve volume
-        updateThickness(hice[i], cice[i], deltaCIce[i], newice[i]);
-        if (deltaCIce[i] < 0) {
-            // Snow is lost if the concentration decreases, and energy is returned to the ocean
-            qow[i] -= deltaCIce[i] * hsnow[i] * Water::Lf * Ice::rhoSnow / tstep.step;
-        } else {
-            // Update snow thickness. Currently no new snow is implemented
-            updateThickness(hsnow[i], cice[i], deltaCIce[i], 0);
-        }
-    }
-}
-
-void IceGrowth::applyLimits(size_t i, const TimestepTime& tstep)
-{
-    if ((0. < cice[i] && cice[i] < IceMinima::c()) || (0. < hice[i] && hice[i] < IceMinima::h())) {
-        qow[i] += cice[i] * Water::Lf * (hice[i] * Ice::rho + hsnow[i] * Ice::rhoSnow) / tstep.step;
-        hice[i] = 0;
-        cice[i] = 0;
-        hsnow[i] = 0;
     }
 }
 } /* namespace Nextsim */
