+++ physics/src/modules/LateralIceSpreadModule/HiblerSpread.cpp
@@ -1,11 +1,14 @@
 /*!
  * @file HiblerSpread.cpp
  *
- * @date 20 Nov 2024
+ * @date 04 Jun 2025
  * @author Tim Spain <timothy.spain@nersc.no>
  */
 
 #include "include/HiblerSpread.hpp"
+
+#include "include/IceMinima.hpp"
+#include "include/constants.hpp"
 
 namespace Nextsim {
 
@@ -34,12 +37,7 @@ ConfigMap HiblerSpread::getConfiguration() const
     };
 }
 
-ModelState HiblerSpread::getStateDiagnostic() const
-{
-    return { { },
-        getConfiguration()
-    };
-}
+ModelState HiblerSpread::getStateDiagnostic() const { return { {}, getConfiguration() }; }
 
 HiblerSpread::HelpMap& HiblerSpread::getHelpText(HelpMap& map, bool getAll)
 {
@@ -71,4 +69,71 @@ void HiblerSpread::melt(const TimestepTime& tstep, double hice, double hsnow, do
     }
 }
 
+void HiblerSpread::newIceFormation(size_t i, const TimestepTime& tst)
+{
+    // Flux cooling the ocean from open water
+    // TODO Add assimilation fluxes here
+    double coolingFlux = qow[i];
+    // Temperature change of the mixed layer during this timestep
+    double deltaTml = -coolingFlux / mixedLayerBulkHeatCapacity[i] * tst.step;
+    // Initial temperature
+    double t0 = sst[i];
+    // Freezing point temperature
+    double tf0 = tf[i];
+    // Final temperature
+    double t1 = t0 + deltaTml;
+
+    // deal with cooling below the freezing point
+    if (t1 < tf0) {
+        // Heat lost cooling the mixed layer to freezing point
+        double sensibleFlux = (tf0 - t0) / deltaTml * coolingFlux;
+        // Any heat beyond that is latent heat forming new ice
+        double latentFlux = coolingFlux - sensibleFlux;
+
+        qow[i] = sensibleFlux;
+        newice[i] = latentFlux * tst.step * (1 - cice[i]) / (Ice::Lf * Ice::rho);
+    } else {
+        newice[i] = 0;
+    }
+}
+
+// Update thickness with concentration
+double HiblerSpread::updateThickness(double& thick, double newConc, double deltaC, double deltaV)
+{
+    return thick += (deltaV - thick * deltaC) / newConc;
+}
+
+void HiblerSpread::lateralIceSpread(size_t i, const TimestepTime& tstep)
+{
+    double deltaCMelt = 0;
+    double deltaCFreeze = 0;
+    freeze(tstep, hice[i], hsnow[i], deltaHi[i], newice[i], cice[i], qow[i], deltaCFreeze);
+    if (deltaHi[i] < 0) {
+        // Note that the cell-averaged hice0 is converted to an ice averaged value
+        melt(tstep, hice0[i], hsnow[i], deltaHi[i], cice[i], qow[i], deltaCMelt);
+    }
+    deltaCIce[i] = deltaCFreeze + deltaCMelt;
+    cice[i] = (hice[i] > 0 || newice[i] > 0) ? cice[i] + deltaCIce[i] : 0;
+    if (cice[i] >= IceMinima::c()) {
+        // The updated ice thickness must conserve volume
+        updateThickness(hice[i], cice[i], deltaCIce[i], newice[i]);
+        if (deltaCIce[i] < 0) {
+            // Snow is lost if the concentration decreases, and energy is returned to the ocean
+            qow[i] -= deltaCIce[i] * hsnow[i] * Water::Lf * Ice::rhoSnow / tstep.step;
+        } else {
+            // Update snow thickness. Currently, no new snow is implemented
+            updateThickness(hsnow[i], cice[i], deltaCIce[i], 0);
+        }
+    }
+}
+
+void HiblerSpread::applyLimits(size_t i, const TimestepTime& tstep)
+{
+    if (cice[i] < IceMinima::c() || hice[i] < IceMinima::h()) {
+        qow[i] += cice[i] * Water::Lf * (hice[i] * Ice::rho + hsnow[i] * Ice::rhoSnow) / tstep.step;
+        hice[i] = 0;
+        cice[i] = 0;
+        hsnow[i] = 0;
+    }
+}
 }
